fixes for previous changes in variables
There were some missing variables that had to be renamed which broke the built. Also updated
squashfs-tools which was previously in the sys-apps category and now it's part of sys-fs. Next
few commits will have kernel config changes and hopefully by end of week we can have a working script.