Updated functions for --co-processor options
On previous commits i removed a case on do_coprocessor function
which checked for certain procs and applied patches as needed,
but this was later removed and it turned out that it was really
needed for only a certain amount of procs except i{5,6}86.
Instead of the case statement I have added a regexp to handle this.